@@ -2,10 +2,10 @@ name: example
 description: An example app for LocalCaptcha package.
 publish_to: none
 
-version: 1.0.1+2
+version: 1.0.2+3
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: '>=3.3.1 <4.0.0'
 
 dependencies:
   flutter:
@@ -16,7 +16,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.0
 
 flutter:
   uses-material-design: true
