@@ -85,11 +85,11 @@ class LocalCaptcha extends StatefulWidget {
   /// Condition for code validation.
   final Duration codeExpireAfter;
 
-  /// Value of captcha when it's generated
+  /// Value of captcha when it's generated.
   final Function(String)? onCaptchaGenerated;
 
   const LocalCaptcha({
-    Key? key,
+    super.key,
     required this.controller,
     this.chars = 'abdefghnryABDEFGHNQRY3468',
     this.length = 5,
@@ -103,8 +103,7 @@ class LocalCaptcha extends StatefulWidget {
     this.codeExpireAfter = const Duration(minutes: 10),
     this.onCaptchaGenerated,
   })  : assert(length > 0),
-        assert(height <= width),
-        super(key: key);
+        assert(height <= width);
 
   @override
   State<LocalCaptcha> createState() => _LocalCaptchaState();
@@ -136,6 +135,7 @@ class _LocalCaptchaState extends State<LocalCaptcha> {
 
       _randomText += String.fromCharCode(charList[index]);
     }
+
     widget.onCaptchaGenerated?.call(_randomText);
   }
 
@@ -144,9 +144,7 @@ class _LocalCaptchaState extends State<LocalCaptcha> {
     super.initState();
 
     widget.controller.setOnValidateFn((code) {
-      if (DateTime.now()
-          .subtract(widget.codeExpireAfter)
-          .isAfter(_lastRefreshAt)) {
+      if (DateTime.now().subtract(widget.codeExpireAfter).isAfter(_lastRefreshAt)) {
         return LocalCaptchaValidation.codeExpired;
       }
 
@@ -225,13 +223,13 @@ class CaptchaTextLayer extends StatelessWidget {
   final List<Color> colors;
 
   const CaptchaTextLayer({
-    Key? key,
+    super.key,
     required this.text,
     this.fontSize,
     required this.height,
     required this.width,
     required this.colors,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -243,9 +241,7 @@ class CaptchaTextLayer extends StatelessWidget {
       alignment: Alignment.center,
       child: Row(
         mainAxisSize: MainAxisSize.min,
-        children: textList
-            .map((e) => _char(String.fromCharCode(e)))
-            .toList(growable: false),
+        children: textList.map((e) => _char(String.fromCharCode(e))).toList(growable: false),
       ),
     );
   }
@@ -255,8 +251,7 @@ class CaptchaTextLayer extends StatelessWidget {
 
     final transform3dPerspective = Matrix4.identity()
       ..setEntry(3, 2, 0.01)
-      ..rotateX(
-          ((random.nextInt(1) + 2) * 0.1) * (random.nextInt(10) >= 5 ? 1 : -1));
+      ..rotateX(((random.nextInt(1) + 2) * 0.1) * (random.nextInt(10) >= 5 ? 1 : -1));
 
     var mFontSize = 0.0;
 
@@ -269,8 +264,7 @@ class CaptchaTextLayer extends StatelessWidget {
       final fontSizeWithTextLength = (autoFontSize * text.length);
 
       if (autoFontSize * text.length > width) {
-        final overflow =
-            (fontSizeWithTextLength - width) / fontSizeWithTextLength * 100;
+        final overflow = (fontSizeWithTextLength - width) / fontSizeWithTextLength * 100;
 
         autoFontSize = height * (fontScale - (fontScale * overflow / 100));
         autoFontSize /= 0.8;
@@ -283,8 +277,7 @@ class CaptchaTextLayer extends StatelessWidget {
       transform: transform3dPerspective,
       alignment: FractionalOffset.center,
       child: Transform.rotate(
-        angle: ((random.nextInt(25) + 5) * pi / 180) *
-            (random.nextInt(10) >= 5 ? 1 : -1),
+        angle: ((random.nextInt(25) + 5) * pi / 180) * (random.nextInt(10) >= 5 ? 1 : -1),
         child: Transform.translate(
           offset: Offset(
             random.nextInt(10) * (random.nextInt(10) >= 5 ? 1 : -1),
@@ -299,9 +292,7 @@ class CaptchaTextLayer extends StatelessWidget {
                 style: TextStyle(
                   color: colors[random.nextInt(colors.length)],
                   fontSize: mFontSize,
-                  fontWeight: (random.nextInt(10) >= 5
-                      ? FontWeight.normal
-                      : FontWeight.bold),
+                  fontWeight: (random.nextInt(10) >= 5 ? FontWeight.normal : FontWeight.bold),
                 ),
               ),
             ),
@@ -318,11 +309,11 @@ class CaptchaNoiseLayer extends StatelessWidget {
   final List<Color> colors;
 
   const CaptchaNoiseLayer({
-    Key? key,
+    super.key,
     required this.height,
     required this.width,
     required this.colors,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
